requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
astroid==2.7.3
attrs==25.1.0
audioread==3.0.1
av==14.1.0
babel==2.17.0
bangla==0.0.2
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
blis==1.2.0
blockdiag==1.5.4
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
bokeh==1.4.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
coqpit==0.0.17
coverage==7.6.12
ctranslate2==4.5.0
cutlet==0.5.0
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dateparser==1.1.8
decorator==5.1.1
diagrams==0.24.4
distlib==0.3.9
docopt==0.6.2
docx==0.2.4
einops==0.8.1
encodec==0.1.1
enum34==1.1.10
faster-whisper==1.1.1
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
fugashi==1.4.0
funcparserlib==2.0.0a0
g2pkk==0.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
graphviz==0.20.3
grpcio==1.70.0
grpcio-status==1.70.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
halo==0.0.31
hangul-romanize==0.1.0
httplib2==0.22.0
huggingface-hub==0.28.1
humanfriendly==10.0
identify==2.6.9
idna==3.10
inflect==7.5.0
isort==5.13.2
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.5
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lazy-object-proxy==1.10.0
librosa==0.10.2.post1
llvmlite==0.44.0
log-symbols==0.0.14
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mccabe==0.6.1
mdurl==0.1.2
mecab-python3==1.0.6
mojimoji==0.0.13
more-itertools==10.6.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.9.1
nodeenv==1.9.1
nose2==0.15.1
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.20.1
openwakeword==0.6.0
packaging==24.2
pandas==1.5.3
pathspec==0.12.1
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.6
pooch==1.8.2
pre_commit==4.2.0
preshed==3.0.9
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pvporcupine==1.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pygame-ce==2.5.3
PyGetWindow==0.0.9
Pygments==2.19.1
pylint==2.10.2
PyMsgBox==1.0.9
pynndescent==0.5.13
pyparsing==3.2.1
pyperclip==1.9.0
pypinyin==0.53.0
PyRect==0.2.0
pysbd==0.3.4
PyScreeze==1.0.1
pyserial==3.5
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-docx==1.1.2
python3-xlib==0.15
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.5.2
schemdraw==0.20
scikit-learn==1.6.1
scipy==1.15.1
setuptools==65.5.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.1
spinners==0.0.24
srsly==2.5.1
srt==3.5.3
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tflite-runtime==2.14.0
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
tornado==6.4.2
tqdm==4.67.1
trainer==0.0.36
transformers==4.48.3
triton==3.2.0
-e /home/pani/api_test/TTS
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
tzlocal==5.3
umap-learn==0.5.7
Unidecode==1.3.8
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.3
vosk==0.3.45
wasabi==1.1.3
weasel==0.4.1
webcolors==24.11.1
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.12.1
yarl==1.18.3

